feat: add a new React component for rendering a 3D force graph with images
This commit adds a new React component called MyGraphComponent, which utilizes the react-force-graph library to render a 3D force graph with images as nodes. The component generates a random connected graph with nodes representing different images. Each node is rendered as a sprite with the corresponding image texture using the THREE.js library. The size and scale of the sprites are also configured within the component.